@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^2.6.1",
     "axios": "^1.8",
     "bootstrap": "^4.5.3",
     "btoa": "^1.2.1",
@@ -12,16 +13,16 @@
     "form-data": "^3.0.0",
     "formik": "^2.1.4",
     "query-string": "^8.1.0",
-    "react": "^17.0.1",
+    "react": "^18",
     "react-bootstrap": "^1.4.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18",
     "react-image-crop": "^8.6.6",
     "react-redux": "^7.1.1",
     "react-router-dom": "^6",
-    "redux": "^4.0.4",
+    "redux": "^5.0.1",
     "redux-thunk": "^2.3.0",
     "socket.io-client": "^4.7.5",
-    "styled-components": "^4.3.2",
+    "styled-components": "^5",
     "typescript": "^5.8.2",
     "yup": "^0.28.3"
   },
@@ -53,9 +54,9 @@
   },
   "devDependencies": {
     "@types/jest": "^26.0.15",
-    "@types/node": "^14.14.8",
-    "@types/react": "^16.9.56",
-    "@types/react-dom": "^16.9.9",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "@types/react-redux": "^7.1.11",
     "@types/react-router-dom": "^5.1.6",
     "@types/react-test-renderer": "^17.0.1",
