@@ -1,34 +1,21 @@
 // import dependencies
 import React from 'react'
-import ReactDOM from 'react-dom'
-import { createStore, applyMiddleware, compose } from 'redux'
+import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
-import thunk from 'redux-thunk'
 import { BrowserRouter } from 'react-router-dom'
 
 // import components and CSS
 import './index.css'
-import reducer from './reducers/rootReducer'
+import { store } from './reducers/rootReducer'
 import App from './App'
 
-declare global {
-  interface Window {
-    __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: typeof compose
-  }
-}
-
-// Redux dev tools
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
-
-// create store and use thunk
-const store = createStore(reducer, composeEnhancers(applyMiddleware(thunk)))
-
 // render App with React Router and Redux
-ReactDOM.render(
+const container = document.getElementById('root')
+const root = createRoot(container)
+root.render(
   <Provider store={store}>
     <BrowserRouter>
       <App />
     </BrowserRouter>
   </Provider>,
-  document.getElementById('root'),
 )
